+++ src/Stratis.SmartContracts.Executor.Reflection/IStateProcessor.cs
@@ -1,0 +1,30 @@
+﻿namespace Stratis.SmartContracts.Executor.Reflection
+{
+    public interface IStateProcessor
+    {
+        /// <summary>
+        /// Applies an externally generated contract creation message to the current state.
+        /// </summary>
+        StateTransitionResult Apply(IState state, ExternalCreateMessage message);
+
+        /// <summary>
+        /// Applies an internally generated contract creation message to the current state.
+        /// </summary>
+        StateTransitionResult Apply(IState state, InternalCreateMessage message);
+
+        /// <summary>
+        /// Applies an internally generated contract method call message to the current state.
+        /// </summary>
+        StateTransitionResult Apply(IState state, InternalCallMessage message);
+
+        /// <summary>
+        /// Applies an externally generated contract method call message to the current state.
+        /// </summary>
+        StateTransitionResult Apply(IState state, ExternalCallMessage message);
+
+        /// <summary>
+        /// Applies an internally generated contract funds transfer message to the current state.
+        /// </summary>
+        StateTransitionResult Apply(IState state, ContractTransferMessage message);
+    }
+}
